Fix provider settings CI review feedback
